02-02 Working Networks

+++ helpers.py
@@ -7,4 +7,4 @@ def quantize(arr, res=0.01, workspace_limits=np.array([[-.6,.6],[-.6,.6],[0,.9]]
     range_ = np.abs(workspace_limits[:,1] - workspace_limits[:,0])
     ndx_range = range_/res 
     ndx = np.round(ndx_range * (arr - workspace_limits[:,0]) / range_)
-    return np.int16(ndx)   
+    return np.int32(ndx)
